// carta-controller-deploy/carta-controller-v2/src/types.ts
export interface CartaLocalAuthConfig {
    publicKeyLocation: string;
    privateKeyLocation: string;
    keyAlgorithm: Algorithm;
    issuer: string;
    refreshTokenAge: string;
    accessTokenAge: string;
    scriptingTokenAge: string;
}
// ...
export enum ScriptingAccess {
    Enabled = "enabled-all-users",
    Disabled = "disabled-all-users",
    OptIn = "opt-in"
}
// ...
export type TokenPayload = {
    username?: string;
    scripting?: boolean;
    iss?: string;
    refresh?: boolean;
    expires_in?: number;
};

// Token verifier function
export type Verifier = (cookieString: string) => TokenPayload | Promise<TokenPayload> | undefined;